Allow real custom options and errors to actually bubble up
Removes the default resolution of 640x360 (why was it set to that in the
first place?).

preserve_aspect_ratio is also something that I wanted to handle myself
with the `custom` calling code.

Using similar ffmpeg options in MediaCrush, here is a valid way for this
gem to be used:

process encode_video: [:mp4,
  custom: "-vcodec libx264 -an -pix_fmt yuv420p -profile:v baseline " +
          "-preset slower -crf 18 " +
          "-vf 'scale=trunc(in_w/2)*2:trunc(in_h/2)*2' -map 0:v:0"]